requirements.txt
streamlit
google-genai
pandas
thefuzz
python-docx
python-dotenv
requests
tabulate
httpx
